Move functional utils to module utils
