+++ bounties_api/std_bounties/tests.py
@@ -6,7 +6,7 @@ from functools import partial
 from django.test import TestCase
 
 # Create your tests here.
-from std_bounties.utils import narrower, formatter, merge, pipe
+from utils.functional_tools import narrower, formatter, merge, pipe
 from std_bounties.models import Bounty, Fulfillment
