+++ bounties_api/utils/functional_tools.py
